Error Checking Integrity
Minor fixes as explained below:

Added another block of else if statement to the username setter
method. This will check for the length of the new username to be
assigned; the method raises a value error if the length of the
new username is less than 3 characters.

Also, spaced out lines of codes to enhance reading