Test Methods
Improved some methods as explained below

tactication module
Did minor fixes to the signup method to test the method's functionality. At the moment,
 the method returns a value of zero to signify that it has successfully
recieved the information passed to it.

tactapp module
Completely developed the auth method which registers a new user. This
method takes inputs from the standard input supplied by th user and does
some error checking. If all is clean, it invokes the signup method from
the tactication module and pass values to it. It prints a welcome message
and terminates the loop if it receives a value of zero, otherwise it
notifies of a registration failure.