Improved method and added new
Improved the signup method of the tactication module and added a new
method as described below

getuuid method
This method generates a varied uuid whenever it is invoked. It first
computes for the length of the supplied username, converts the generated
uuid to hexadecimal and appends slices of the username to the generated
uuid. It then returns a string representation of the id to the caller.

This method is invoked in the signup() method which is used to generate
a userid