Debugged login issue
After several tries, I've managed to debug the login issue. This
problem, for some reason I'm yet to understand has to do with the naming
of the database table.

Initially, I used the uuid generated for the new user as the name of the
table. Although committing data to the table executed without errors,
fetching from it was the main challenge. Instead of returning the query
fetch results, I decided to print it to the console only to see an empty
list; thus the query was unable to fetch the data.

So I tried another approach; I created a new table with a simple name,
users and updated the various methods in the associated modules that
depend on it. I ran the program once more and viola, the
query results printed the fetched data to the console as a list of
a tuple.

Although this program will not be used in the production environment,
I've learned a key thing about database table names. Also, I discovered
how to use the time module to manage the function calls and I'm looking
forward to learn more.