Updated tactuser module
Updated the tactuser module to invoke two methods from the dbmanager
module depending the value passed to it as parameter. If the value in
the parameter is 'add', the commit method invokes the save_record from
the dbmanager module to save a new record.

However, if the value in the parameter is 'dit', the commit method
invokes the edit_record method from the dbmanager module to update the
data of an existing record.