Method improvement
Currently working on the edit_record method of the dbmanager module. The
method does what I want it to do but I think there is more improvement
to do. The following explains everything.

The method takes a list as a parameter. The first element of the list is
an option indicating what the user wants to do: either update all data
of a row, a single data of a row or a selection of data of a row. The
second option is the id of the data to update its record and the other
options are the values to update the record. The current code is
redundant from unnecessary repeatitions.

Several ececutions updated the database